requirements.txt
amqp==5.1.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
bcrypt==4.0.1
billiard==4.1.0
boto3==1.33.7
botocore==1.33.7
cachecontrol==0.13.1
cachetools==5.3.1
celery==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cryptography==41.0.3
decorator==4.4.2
distro==1.8.0
Django==3.2
django-codemirror2==0.2
django-cors-headers==3.7.0
django-cte==1.1.4
django-js-asset==1.2.2
django-mptt==0.9.1
django-multitenant==4.0.0
django-simple-history==3.3.0
django-storages==1.14.2
djangorestframework==3.14.0
dockerpty==0.4.1
docopt==0.6.2
drf-yasg==1.21.7
environs==9.5.0
et-xmlfile==1.1.0
firebase-admin==6.2.0
google-api-core==2.11.1
google-api-python-client==2.99.0
google-auth==2.23.0
google-auth-httplib2==0.1.1
google-cloud-core==2.3.3
google-cloud-firestore==2.12.0
google-cloud-storage==2.10.0
google-crc32c==1.5.0
google-resumable-media==2.6.0
googleapis-common-protos==1.60.0
grpcio==1.58.0
grpcio-status==1.58.0
httplib2==0.22.0
httpx==0.27.0
idna==3.4
inflection==0.5.1
jmespath==1.0.1
jsonschema==3.2.0
kombu==5.3.1
lxml==4.9.3
marshmallow==3.20.1
mock==5.1.0
msgpack==1.0.5
numpy==1.25.2
openpyxl==3.1.2
packaging==23.1
pandas==2.1.0
paramiko==3.3.1
pdfkit==1.0.0
Pillow==10.0.0
plivo==4.40.0
prompt-toolkit==3.0.39
proto-plus==1.22.3
protobuf==4.24.3
psycopg==3.1.10
psycopg2-binary==2.9.9
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.8.0
PyNaCl==1.5.0
pyotp==2.9.0
pyparsing==3.1.1
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==0.21.1
pytz==2023.3.post1
PyYAML==5.4.1
redis==5.0.0
reportlab==4.0.4
requests==2.31.0
rsa==4.9
s3transfer==0.8.2
six==1.16.0
sqlparse==0.4.4
texttable==1.6.7
typing-extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.16
vine==5.0.0
wcwidth==0.2.6
websocket-client==0.59.0
XlsxWriter==3.1.3
